Allow the use of a legend stored in the session.
Session can be used to store multiple legends and these legend
definition are similar to Legend objects json representation so they
allow the same flexibility.

Wether the session or the database entries are used depend on the store
parameter. By default, or if none is profided, the database store is
used.

Shortcuts functions defined in raster.shrotcuts provide a wrapper to set
and get colormap from session.